Nanoparticle Systems: Interaction Potential Computation
A collection of Python scripts for computing the interaction potential
energy profiles in nanoparticle systems including, but not limited to,
nanoparticle-electrode ensembles. The two principle forces accounted
for in the computations are the Coulombic interactions between
dissociated charge-bearing ligands functionalising the surface of the
nanoparticles in addition to the intrinsic van der Waals forces between
the metal cores.